refactor navigation into client components
