+++ src/app/constants.ts
@@ -1,0 +1,18 @@
+export const NAV_ITEMS = [
+  { text: '首页', link: 'http://www.csrc.gov.cn/csrc/index.shtml' },
+  { text: '机构概况', link: 'http://www.csrc.gov.cn/csrc/jggk/index.shtml' },
+  { text: '新闻发布', link: 'http://www.csrc.gov.cn/csrc/xwfb/index.shtml' },
+  { text: '政务信息', link: 'http://www.csrc.gov.cn/csrc/zwxx/index.shtml' },
+  { text: '办事服务', link: 'http://www.csrc.gov.cn/csrc/bsfw/index.shtml' },
+  { text: '互动交流', link: 'http://www.csrc.gov.cn/csrc/hdjl/index.shtml' },
+  { text: '统计信息', link: 'http://www.csrc.gov.cn/csrc/tjsj/index.shtml' },
+  { text: '专题专栏', link: 'http://www.csrc.gov.cn/csrc/ztzl/index.shtml' }
+];
+
+export const LANGUAGES = [
+  'ENGLISH', '繁體中文', 'PORTUGUÊS (BRASIL)', 'POLSKI', '한국어', '日本語',
+  'ITALIANO', 'FRANÇAIS', 'ESPAÑOL (MÉXICO)', 'ESPAÑOL', 'DEUTSCH', 'РУССКИЙ'
+];
+
+export const BRAND_COLOR = '#D00403';
+export const FONT_FAMILY = 'Helvetica, Arial, sans-serif';
